added loader and some refactor changes
